Code comments.

+++ Pizzapalatset/ViewModel/OrderViewModel.cs
@@ -17,20 +17,22 @@ namespace Pizzapalatset.ViewModel
     {
         public ObservableCollection<Pizza> OrderList = new ObservableCollection<Pizza>();
         public ObservableCollection<Order> httpOrderList = new ObservableCollection<Order>();
-
+        
         HttpClient httpClient;
 
         string orderUrl = "https://localhost:44360/api/orders/";
         string pOrderUrl = "https://localhost:44360/api/pizzaorders/";
         public Order MyOrder { get; set; }
-
         public OrderViewModel()
         {
             MyOrder = new Order();
             httpClient = new HttpClient();
             httpOrderList = new ObservableCollection<Order>();
         }
-
+        /// <summary>
+        /// Method that adds a pizza to the OrderList and calculates the total cost of the order.
+        /// </summary>
+        /// <param name="pizza"></param>
         public void AddToCart(Pizza pizza)
         {
             if (pizza != null)
@@ -39,6 +41,10 @@ namespace Pizzapalatset.ViewModel
                 CalculateTotalCost(pizza, true);
             }
         }
+        /// <summary>
+        /// Method that removes a pizza to the OrderList and recalculates the total cost of the order.
+        /// </summary>
+        /// <param name="pizza"></param>
         public void RemoveFromCart(Pizza pizza)
         {
             if (pizza != null)
@@ -47,6 +53,9 @@ namespace Pizzapalatset.ViewModel
                 CalculateTotalCost(pizza, false);
             }
         }
+        /// <summary>
+        /// Method that deletes all the elements in the OrderList in case the user wants to cancel their order.
+        /// </summary>
         public async void CancelOrder()
         {
             var cancel = new MessageDialog("Är du säker på att du vill avbryta din order?", "Avbryt order");
@@ -65,7 +74,12 @@ namespace Pizzapalatset.ViewModel
                 MyOrder.TotalCost = 0;
             }
         }
-
+        /// <summary>
+        /// Method that verifies if the user wants to proceed with their order. It also checks if the order is empty, 
+        /// in that case the user is prompted to add pizzas to the order.
+        /// </summary>
+        /// <param name="o"></param>
+        /// <returns></returns>
         public async Task ConfirmOrder(Order o)
         {
             var ready = new MessageDialog("Är du redo att skicka din beställning till köket?", "Bekräfta order");
@@ -80,7 +94,7 @@ namespace Pizzapalatset.ViewModel
             {
                 MessageDialog confirm = new MessageDialog($"Din order har skickats till köket. Smaklig måltid!");
                 await confirm.ShowAsync();
-                await CreateOrderinDB(o);
+                await PostOrderToDB(o);
                 OrderList.Clear();
                 MyOrder.TotalCost = 0;
             }
@@ -90,6 +104,12 @@ namespace Pizzapalatset.ViewModel
                 await empty.ShowAsync();
             }
         }
+        /// <summary>
+        /// Method that calculates the total cost of the order, depending on the Add/Remove methods.
+        /// </summary>
+        /// <param name="p"></param>
+        /// <param name="tf"></param>
+        /// <returns></returns>
         public int CalculateTotalCost(Pizza p, bool tf)
         {
             if (tf)
@@ -104,12 +124,12 @@ namespace Pizzapalatset.ViewModel
         }
 
         /// <summary>
-        /// Creates and posts an order to the order table + post to the junction table. 
+        /// Creates and posts an order to the order table + post to the junction table in the server database.
         /// </summary>
         /// <param name="o"></param>
         /// <returns></returns>
         /// 
-        public async Task CreateOrderinDB(Order o)
+        public async Task PostOrderToDB(Order o)
         {
             var myOrder = JsonConvert.SerializeObject(o);
             HttpContent httpContent = new StringContent(myOrder);
@@ -135,15 +155,32 @@ namespace Pizzapalatset.ViewModel
             MessageDialog msg = new MessageDialog($"Du har fått ordernummer: {orderNo}");
             await msg.ShowAsync();
         }
-
-        public async Task DeleteOrderAsync(int id)
+        /// <summary>
+        /// In case a user wants to cancel an order, they can type the ordernumber to do that.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
+        public async Task DeleteOrderAsync(string value)
         {
-            var httpClient = new HttpClient();
-            var result = await httpClient.DeleteAsync(orderUrl + id);
-            MessageDialog msg = new MessageDialog("Din order är avbruten.", $"Order {id}");
-            await msg.ShowAsync();
-        }
+            bool check = Int32.TryParse(value, out int id);
 
+            if(check)
+            {
+                var httpClient = new HttpClient();
+                var result = await httpClient.DeleteAsync(orderUrl + id);
+                MessageDialog msg = new MessageDialog("Din order är avbruten.", $"Order {id}");
+                await msg.ShowAsync();
+            }
+            else
+            {
+                MessageDialog msg = new MessageDialog("Ordern kunde inte hittas, vänligen kolla så att ordernumret är korrekt");
+                await msg.ShowAsync();
+            }
+        }
+        /// <summary>
+        /// Method that retrieves all the current orders in the database.
+        /// </summary>
+        /// <returns></returns>
         public async Task<ObservableCollection<Order>> GetOrdersAsync()
         {
             var jsonProducts = await httpClient.GetStringAsync(orderUrl);
